refactor(core): split mathematics extension into focused modules
Extract the two pure, DOM-free concerns from the 740-line mathematics
extension into co-located modules:

- mathematics-render.ts: the katex lazy loader and renderKatex.
- mathematics-markdown.ts: the input-rule regexes and the inline / block
  markdown-it rules.

mathematics.ts keeps the two node specs, the input-rule factory, and the
extension factory, and re-exports loadKatex so the module export surface
is unchanged.
